Rudimentary file browsing support
Implemented a special new FilesystemAlbum.java for managing the
"filesystem" special album. I wanted to use nested submediasets, but the
gallery was too stupid for that and wanted to scan and join all of the
submediasets into a single "master" mediaset, which is really bad.

Instead, AlbumPage.java is hacked to recognize special MediaObject's and
reload its own contents when one such object is selected. These special
MediaObjects represent directories within the filesystem.

There are currently three problems in the implementation;
1) There is no "video" support,
2) The UriImage MediaObject seems to remain open, so loading directories
with LOTS of images will eventually crash with a "Too many open files"
error,
3) There is no *LABEL* on the directories presented within the
AlbumPage.