Add new classes for local filesystem browsing
FilesystemMediaItem is the superclass for FilesystemImage and (not yet
implemented) FilesystemVideo. There are some things in FilesystemImage
that should be moved to FilesystemMediaItem.

This solves the too many open files problem caused by use of UriImage.

This is based on LocalImage and LocalMediaItem which are used for
browsing media stored in the media database.